requirements.txt
pydantic==2.8.2
PyYAML==6.0.2
opencv-python==4.10.0.84
scipy==1.14.0